rename testWorldIT
